+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 * Add a visual cue to invalid entry on transformation form. UIDEXP-220.
 * Update `stripes` to `v6.0.0`. UIDEXP-221.
 * Clear the validation error after user adjusts data entry in transformation field. UIDEXP-218.
+* Add placeholder to the transformation form. UIDEXP-219.
 
 ## [3.0.2](https://github.com/folio-org/ui-data-export/tree/v3.0.2) (2020-11-13)
 [Full Changelog](https://github.com/folio-org/ui-data-export/tree/v3.0.1...v3.0.2)